# app/Backend/kratos/spec/swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome to the Ory Kratos HTTP API documentation!",
    "title": "Ory Kratos",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
...
    "/self-service/settings": {
      "post": {
        "security": [
          {
            "sessionToken": []
          }
        ],
        "description": "Use this endpoint to complete a settings flow by sending an identity's updated password. This endpoint\nbehaves differently for API and browser flows.\n\nAPI-initiated flows expect `application/json` to be sent in the body and respond with\nHTTP 200 and an application/json body with the session token on success;\nHTTP 303 redirect to a fresh settings flow if the original flow expired with the appropriate error messages set;\nHTTP 400 on form validation errors.\nHTTP 401 when the endpoint is called without a valid session token.\nHTTP 403 when `selfservice.flows.settings.privileged_session_max_age` was reached or the session's AAL is too low.\nImplies that the user needs to re-authenticate.\n\nBrowser flows without HTTP Header `Accept` or with `Accept: text/*` respond with\na HTTP 303 redirect to the post/after settings URL or the `return_to` value if it was set and if the flow succeeded;\na HTTP 303 redirect to the Settings UI URL with the flow ID containing the validation errors otherwise.\na HTTP 303 redirect to the login endpoint when `selfservice.flows.settings.privileged_session_max_age` was reached or the session's AAL is too low.\n\nBrowser flows with HTTP Header `Accept: application/json` respond with\nHTTP 200 and a application/json body with the signed in identity and a `Set-Cookie` header on success;\nHTTP 303 redirect to a fresh login flow if the original flow expired with the appropriate error messages set;\nHTTP 401 when the endpoint is called without a valid session cookie.\nHTTP 403 when the page is accessed without a session cookie or the session's AAL is too low.\nHTTP 400 on form validation errors.\n\nDepending on your configuration this endpoint might return a 403 error if the session has a lower Authenticator\nAssurance Level (AAL) than is possible for the identity. This can happen if the identity has password + webauthn\ncredentials (which would result in AAL2) but the session has only AAL1. If this error occurs, ask the user\nto sign in with the second factor (happens automatically for server-side browser flows) or change the configuration.\n\nIf this endpoint is called with a `Accept: application/json` HTTP header, the response contains the flow without a redirect. In the\ncase of an error, the `error.id` of the JSON response body can be one of:\n\n`session_refresh_required`: The identity requested to change something that needs a privileged session. Redirect\nthe identity to the login init endpoint with query parameters `?refresh=true\u0026return_to=\u003cthe-current-browser-url\u003e`,\nor initiate a refresh login flow otherwise.\n`security_csrf_violation`: Unable to fetch the flow because a CSRF violation occurred.\n`session_inactive`: No Ory Session was found - sign in a user first.\n`security_identity_mismatch`: The flow was interrupted with `session_refresh_required` but apparently some other\nidentity logged in instead.\n`security_identity_mismatch`: The requested `?return_to` address is not allowed to be used. Adjust this in the configuration!\n`browser_location_change_required`: Usually sent when an AJAX request indicates that the browser needs to open a specific URL.\nMost likely used in Social Sign In flows.\n\nMore information can be found at [Ory Kratos User Settings \u0026 Profile Management Documentation](../self-service/flows/user-settings).",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "frontend"
        ],
        "summary": "Complete Settings Flow",
        "operationId": "updateSettingsFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Settings Flow ID\n\nThe value for this parameter comes from `flow` URL Query parameter sent to your\napplication (e.g. `/settings?flow=abcde`).",
            "name": "flow",
            "in": "query",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateSettingsFlowBody"
            }
          },
          {
            "type": "string",
            "description": "The Session Token of the Identity performing the settings flow.",
            "name": "X-Session-Token",
            "in": "header"
          },
          {
            "type": "string",
            "description": "HTTP Cookies\n\nWhen using the SDK in a browser app, on the server side you must include the HTTP Cookie Header\nsent by the client to your server here. This ensures that CSRF and session cookies are respected.",
            "name": "Cookie",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "settingsFlow",
            "schema": {
              "$ref": "#/definitions/settingsFlow"
            }
          },
          "303": {
            "$ref": "#/responses/emptyResponse"
          },
          "400": {
            "description": "settingsFlow",
            "schema": {
              "$ref": "#/definitions/settingsFlow"
            }
          },
          "401": {
            "description": "errorGeneric",
            "schema": {
              "$ref": "#/definitions/errorGeneric"
            }
          },
          "403": {
            "description": "errorGeneric",
            "schema": {
              "$ref": "#/definitions/errorGeneric"
            }
          },
          "410": {
            "description": "errorGeneric",
            "schema": {
              "$ref": "#/definitions/errorGeneric"
            }
          },
          "422": {
            "description": "errorBrowserLocationChangeRequired",
            "schema": {
              "$ref": "#/definitions/errorBrowserLocationChangeRequired"
            }
          },
          "default": {
            "description": "errorGeneric",
            "schema": {
              "$ref": "#/definitions/errorGeneric"
            }
          }
        }
      }
    },
    "/self-service/settings/api": {
      "get": {
        "description": "This endpoint initiates a settings flow for API clients such as mobile devices, smart TVs, and so on.\nYou must provide a valid Ory Kratos Session Token for this endpoint to respond with HTTP 200 OK.\n\nTo fetch an existing settings flow call `/self-service/settings/flows?flow=\u003cflow_id\u003e`.\n\nYou MUST NOT use this endpoint in client-side (Single Page Apps, ReactJS, AngularJS) nor server-side (Java Server\nPages, NodeJS, PHP, Golang, ...) browser applications. Using this endpoint in these applications will make\nyou vulnerable to a variety of CSRF attacks.\n\nDepending on your configuration this endpoint might return a 403 error if the session has a lower Authenticator\nAssurance Level (AAL) than is possible for the identity. This can happen if the identity has password + webauthn\ncredentials (which would result in AAL2) but the session has only AAL1. If this error occurs, ask the user\nto sign in with the second factor or change the configuration.\n\nIn the case of an error, the `error.id` of the JSON response body can be one of:\n\n`security_csrf_violation`: Unable to fetch the flow because a CSRF violation occurred.\n`session_inactive`: No Ory Session was found - sign in a user first.\n\nThis endpoint MUST ONLY be used in scenarios such as native mobile apps (React Native, Objective C, Swift, Java, ...).\n\nMore information can be found at [Ory Kratos User Settings \u0026 Profile Management Documentation](../self-service/flows/user-settings).",
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "frontend"
        ],
        "summary": "Create Settings Flow for Native Apps",
        "operationId": "createNativeSettingsFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Session Token of the Identity performing the settings flow.",
            "name": "X-Session-Token",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "settingsFlow",
            "schema": {
              "$ref": "#/definitions/settingsFlow"
            }
          },
          "400": {
            "description": "errorGeneric",
            "schema": {
              "$ref": "#/definitions/errorGeneric"
            }
          },
          "default": {
            "description": "errorGeneric",
            "schema": {
              "$ref": "#/definitions/errorGeneric"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "errorGeneric": {
      "description": "The standard Ory JSON API error format.",
      "type": "object",
      "title": "JSON API Error Response",
      "required": [
        "error"
      ],
      "properties": {
        "error": {
          "$ref": "#/definitions/genericError"
        }
      }
    },
...
    "settingsFlow": {
      "description": "This flow is used when an identity wants to update settings\n(e.g. profile data, passwords, ...) in a selfservice manner.\n\nWe recommend reading the [User Settings Documentation](../self-service/flows/user-settings)",
      "type": "object",
      "title": "Flow represents a Settings Flow",
      "required": [
        "id",
        "type",
        "expires_at",
        "issued_at",
        "request_url",
        "ui",
        "identity",
        "state"
      ],
      "properties": {
        "active": {
          "description": "Active, if set, contains the registration method that is being used. It is initially\nnot set.",
          "type": "string"
        },
        "expires_at": {
          "description": "ExpiresAt is the time (UTC) when the flow expires. If the user still wishes to update the setting,\na new flow has to be initiated.",
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "description": "ID represents the flow's unique ID. When performing the settings flow, this\nrepresents the id in the settings ui's query parameter: http://\u003cselfservice.flows.settings.ui_url\u003e?flow=\u003cid\u003e",
          "type": "string",
          "format": "uuid"
        },
        "identity": {
          "$ref": "#/definitions/identity"
        },
        "issued_at": {
          "description": "IssuedAt is the time (UTC) when the flow occurred.",
          "type": "string",
          "format": "date-time"
        },
        "request_url": {
          "description": "RequestURL is the initial URL that was requested from Ory Kratos. It can be used\nto forward information contained in the URL's path or query for example.",
          "type": "string"
        },
        "return_to": {
          "description": "ReturnTo contains the requested return_to URL.",
          "type": "string"
        },
        "state": {
          "$ref": "#/definitions/settingsFlowState"
        },
        "type": {
          "$ref": "#/definitions/selfServiceFlowType"
        },
        "ui": {
          "$ref": "#/definitions/uiContainer"
        }
      }
    },
...
    "updateSettingsFlowBody": {
      "description": "Update Settings Flow Request Body",
      "type": "object"
    },
...
  },
  "responses": {
    "emptyResponse": {
      "description": "Empty responses are sent when, for example, resources are deleted. The HTTP status code for empty responses is typically 201."
    },
...
  },
  "securityDefinitions": {
    "oryAccessToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}